Move some tests inside lib.rs to keep function private
